+++ flight-service/src/main/java/com/oneul_tanda/flight_service/infrastructure/config/querydsl/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package com.oneul_tanda.flight_service.infrastructure.config;
+package com.oneul_tanda.flight_service.infrastructure.config.querydsl;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
